fix: 구독자 수 데이터 타입 변경

+++ src/main/resources/schema.sql
@@ -69,7 +69,7 @@ CREATE TABLE interests
 (
     "id"               UUID PRIMARY KEY,
     "name"             VARCHAR(50)       NOT NULL,
-    "subscriber_count" INTEGER DEFAULT 0 NOT NULL,
+    "subscriber_count" BIGINT DEFAULT 0 NOT NULL,
     "created_at"       TIMESTAMPTZ       NOT NULL,
     "updated_at"       TIMESTAMPTZ       NULL
 );
